+++ src/mask.c
@@ -29,3 +29,4 @@ U64 blackPassedMasks[64];
 
 
 
+
