+++ src/mask.h
@@ -241,7 +241,7 @@ inline U64 setFileRankMask(int file_number, int rank_number) {
 }
 
 
-extern void initEvaluationMasks() {
+void initEvaluationMasks() {
 
     // loop over ranks
     for (int rank = 0; rank < 8; rank++)
