+++ src/zobrist.c
@@ -7,9 +7,4 @@
 
 
 
-// random piece keys [piece][square]
-U64 pieceKeys[12][64];
-// random enpassant keys [square]
-U64 enpassantKeys[64];
-// random castling keys
-U64 castleKeys[16];
+
