+++ src/zobrist.h
@@ -6,10 +6,10 @@
 
 
 // random piece keys [piece][square]
-extern U64 pieceKeys[12][64];
+U64 pieceKeys[12][64];
 // random enpassant keys [square]
-extern U64 enpassantKeys[64];
+U64 enpassantKeys[64];
 // random castling keys
-extern U64 castleKeys[16];
+U64 castleKeys[16];
 
 extern void initRandomKeys();
